stage2_baseline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fly-olf-stage2-baseline"
version = "0.1.0"
description = "Stage 2 baseline modeling for fly olfactory learning"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "door-python-toolkit>=0.1.0",
]

[project.scripts]
fly-olf-stage2 = "fly_olf.stage2_baseline.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
